test(mdxish): cover all four bounded regex families in the bounds pins
CodeRabbit review follow-up: the long-word and long-whitespace
output-equivalence pins only exercised **, but the prefix bound changed
all four regex families, each with its own lookaheads and content
clauses. Parameterize both pins across **, __, *, and _.

Also moves the pathological-performance / bounds describe into its own
test file so the behavior-matrix file stays under a thousand lines.
